Added code to define, allocate, and initialize the CRF
WARNING: Current implementation consumes over 32 GB of RAM and
can not run on my system for the testing data I was using which
had about 1200 individuals on 65000 SNPs. Some rethinking needs
to be done to reduce memory usage.

Also modified Inputline.cpp to specify the --regions option to
bcftools if bcftools is being used to read the file, allowing it
to skip straight to the chromosome of interest and speeding things
up by a huge amount when the chromosome selected is far into the
file or at the end. It will probably cause a bcftools error though
if the file is not indexed. The program does not handle such an
error itself. We must verify that the bcftools error is printed
on the screen if this happens.