Modified major arrays to reduce memory usage dramatically
Came up with a scheme to encode a 0.0 to 1.0 value as an int8_t,
effectively defining 8 bit floats using primitive operations and
macros. Also rechunked memory allocation to a single chunk for
all windows and pointers multiplied out on the base 0 window which
is allocated memory for all windows at once. This could be reduced
further by using multiplication on access to index a single large
array, eliminating the 64 bit pointers to each window within the
large chunk. Considering this further, I do not think the memory
savings here is worth introducing the complexity to the code.

This majorly reduced memory usage roughly by a factor of 6 to 8,
allowing the program to run with about 18 GB of RAM for chromosome
2 of the hapmap test data I am playing with and a CRF spacing of
1 SNP. There are about 110,000 SNPs on chromosome 2, and 1254
individuals. The number of subpops is also a factor, and currently
I don't have them aggregated and the program detects all phase 3
sample groups as seperate subpops.