Pulled in mm.[ch] and converted to a C++ class (was a C-style object)
This is a class that creates an object that allocates memory internally
in large chunks (here using C memory allocation, but it doesn't matter)
and a user of the class to call an allocate function that returns
memory parceled out of the large chunks. The intention is for this
to be used in tightly threaded code where the code needs to allocate
lots of very small chunks but can not know in advance how many. The
internal buffer allocation in large chunks makes calls to malloc()
rare rather than very frequent, preventing excessive thread contention
for the heap memory allocator. The individual allocations can not be
free selectively and individually. The intended use is where the
calling thread to allocate() reaches a point in its execution where
it knows that all of the many small objects allocated through the
class can be freed all at once. This is accomplished by calling
recycle(), which resets the internal pointers to the beginning of the
buffers allocated, or the destructor which will internally free all
chunks as well as the object and render it invalid for further use.