Defined some data structures and added window setup code to random-forest.cpp
Have not tested compilation or execution. None of the actual tree part
is implemented yet. Still working on the code that is above and flanking
the tree building and evaluation stages.

Decided on explicitly building trees in actual tree structures rather
than implicitly building them on the recursive call stack as RFMIX
version 1 does. The trees will be built by fully classifying the
reference data, then the query samples will be evaluated by these
trees. The point of this is to allow future expansion/development
into pre-training trees and storing them in a file, then using these
pre-trained trees to run the program without an explicit set of
reference haplotypes, just the tree file. Actual implementation of
this is deferred in planned version 2 development. This will likely
be a version 3 feature, but hopefully won't require a full rewrite
of the code.