Added option to filter on amount of missing data, and capability of filtering on maf.
Filtering on amount of missing data is available now as a command
line option. Filtering on maf or mac, is possible now but is not
(yet) a command line option. Note that filtering on either for
the admixed/query panel results in results that change if some of
the samples are run again but with different additional samples
run, or if some or one of the samples are run alone without the
others.