Major debugging in random-forest.cpp, and addition of tree evaluation code.
Found several bugs, one of them a really big problem where the SNP
marked at the node was not the same one as the SNP selected for
dividing the haplotypes. This resulted in tree evaluation being
completely off. It is fixed. Current implementation produces very
strong (like 0.99) assignments for the testdata set that was provided
with RFMIX v1, but weaker on the clinical cases dataset used recently
with RFMIX v1. Some further troubleshooting might be needed here.