Added forward-backward algorithm
In the process, converted sample->current_p[2][][] to remove a
dimension that adds excessive memory for 8-byte pointers for
every CRF window, when as int8_t the values for each subpop for
each of these CRF windows was only 4 bytes with 4 subpops. The
array is now sample->current_p[2][ IDX(w,s) ] where the IDX
macro calculates w*n_subpops + s, w being the window index and
s being the subpop index.

Also converted sample->current_p to int16_t and added DF16 and EF16
macros. This is for more precision for the results produced by
the forward-backward algorithm.