Fixed inane n_subpops-1 requirement throughout the code.
query/admixed samples are now marked -1 as apriori subpop, and
have no reference population string in the mapping for that. Any
output code will have to explicitly check for -1 instead of rely
on the reference population string mapping to output an appropriate
token.

Much better this way and much less error prone or confusing
throughout random-forest.cpp and crf.cpp.