Created inline functions for ef8 and ef16, add recombination to viterbi
Well, there was a recombination state transition factor there already
but it was mostly a place holder and purposely incorrect before. Now
it is fixed and using the same thing as forward-backward. It could
be precomputed in an array and the same values used for both viterbi
and forward-backward from that array, and also for all samples.