Started adding code to perform admixture simulations loading from VCF
Wanted to reuse VCF loading code from rfmix but I really did not
code it for such purposes, so I dropped out some new code. If I
ever have the time, I would like to make handling of VCF into a
class and cleanup the mess in load-input.cpp to use it, and have it
used by the simulate program as well.

The simulate program does nothing at the moment except parse its
arguments and load the VCF input.