Added SNP alleles & output of them to simulator
To merge simulated VCF output with samples from the original
VCF input or other real samples defined at the same SNPs, I
needed the the simulator to output the alleles for SNPs, and
snp ids, as they were given
in the VCF input rather than just output A and C as the alleles
and just . for the snp id. This is for creating merged VCFs to
use in generation of RFMIX v1 input for comparison purposes.