Changed est_p to int16_t, several other small changes.
Trying to track down why this program is consistently slightly
worse than v1, and on less proper simulated data it was worse
enough that we care. None of these changes resulted in major
improvements though.

Also, making the CRF smaller with RF the same makes the results
worse not better by either mean accuracy or variance. One of the
major points of this code is this feature which v1 could not
accomodate, but it is proving counter-intuitive in its effects. Do
I need a conditional dependence term due to overlapping RF?