Reworked window layouts
Modified and redesigned how the CRF window positions are
calculated and what SNPs will be included in its RF windows for
tree building. Results got slightly worse, but the code is
improved. Deciding to keep it.

Also added an option --analyze-range=<start pos>-<end pos> to
restrict analysis to the specified range of the chromosome, in Mbp.
Unfortunately, it is not useful for speeding up simulated data
trials, because the output is then a subset of the verification
data and misaligned. Not easy to fix that, but it is only a
problem if attempting to use it to speed up debugging runs.
The solution is just to simulate on a smaller segment of a chromosome
that you extract with bcftools first.

Also added the --minimum-snps=<#> option which forces RF windows
to have at least this many SNPs, even if the genetic size of the
window will exceed the RF window size.