Fixed sample map loading if samples not present in ref vcf
Previously every non-commented-out line in the sample map file
would create a sample record in the main data structure. However,
some of these samples may not be present in the reference VCF
given. This is especially likely if someone uses a common master
sample map for a large reference panel, but then only uses subsets
of it as the reference VCF. Indeed this was the case when I was
trying to do simulations and the parents of my simulated individuals
had been removed from the reference VCF, but they were still listed
in the sample map. This can create all kinds of havoc because
haplotypes are not found and loaded for the reference sample
records created, yet the program will still run normally. The
output can have unexpected results because of subpop <-> sample
mapping differences that would not occur had the non-existant
reference samples in the sample map simply be ignored.

Now, samples defined in the sample map that were not found in the
reference VCF are ignored and no sample record created.

Just don't worry about it if none of that makes sense. The
problem is no longer a problem.