Modifications to CRF viterbi and fb for weighting
Adding a weight term on the state-change probability that effectively
accounts for the overlap of windows and denser CRF. How to know
what to set this value to is unclear. Testing on hapmap3 simulated
data suggests weights of 4-8 for a CRF of 50 SNPs and RF of 100 SNPs.
Higher values, like 32 or 64 for a CRF of 12 SNPs and RF of 100 SNPs.
It is not clear to me whether the value will depend on the reference
panel composition and SNP density, or whether it is solely a
function of the window overlap. Suyash says we need to "learn" its
value through simulations. I guess by simulations we will empirically
determine values and perhaps an ad-hoc formula for computing a value
in the program automatically based on the amount of window overlap.

Adding this parameter, and setting it to "good" values, results in
excellent performance on HapMap3 simulated data, reaching 98%
accuracy and exceeding the performance of rfmix v1 on the same
data.