Minimum effective genetic distance in CRF
Needed to prevent some sort of underflow causing forward-backward
results to sometimes be all zeros. I have yet to figure out exactly
where this is happening and how. It seems to happen if the
genetic distance between two crf points is very very small. I
force the value now to 0.01 cM regardless of how close the crf
points are, and the problem seems to be mitigated. In this
commit I also have switched some normalizations of probability
vectors where values may be very small and far apart, to a
log normalization with a constant added before exponentiation, so
the largest of the probability values will be 1 in the result
rather than everything flush to zero. This address some of the
problem and mitigates some instances, but not all computations
can be done this way. This needs some more attention.