Rolling up version v2.01. Add first attempts at a manual in Markdown
Also added a perl script to score MSP results rather than rely on
RFMIX v1's R script. The perl script is faster and less clunky
(but still pretty clunky)

Added functionality for --reanalyze-reference which before did
nothing and the reference was always reanalyzed by EM. Now, it is
not by default, but this option will enable it.