Updates to internal simulation, settling of internal parameters
This version seems to work well and repeatably. Constants governing
internal simulation are in rfmix.h. The maximum simulated samples
is now set at 200 diploid samples for each input subpop, the number
of parents is 1/10th of each subpop, and the growth rate is 20% per
generation. Just leave it alone...

CRF weight is scanned from 1 to 100, but stops when the simulation
results are less than half as good as the best weight so far.
Overall the weight curve is simple and convex, but near its
putative maximum it is rather flat and not entirely smooth, but
the variation is just stochastic noise. Probably any of the weights
in the region would give very similar results.