Prevent total loss by genetic drift of a subpopulation during CRF optimization simulation
Each generation of simulation, the code will check for excessive loss of
one or more subpopulations from the progeny that was generated, and
repeat creating the generation (with different random numbers) if any
one population is near totally lost.

An additional countermeasure is the constant simulation growth rate
of 1.2 (currently) which is a #define constant in rfmix.h, is
overridden and set to 2.0 (double the population) if the current
population size is less than 100 (arbitrarily chosen). This helps
when reference panels are small in general, while the above deals
mainly with the case where one or more subpopulations can contribute
only a single individual to found the internal simulation admixed
population.

The effect of a total loss of one or more subpopulations causes
a numerical singularity in attempting to find the optimal value
of the CRF weighting parameter, which propogates until it
ultimately causes a segfault.

 Please enter the commit message for your changes. Lines starting